refactor SDK and add types to URL
