+++ Makefile
@@ -5,6 +5,7 @@ target/release/spinjs: target/wasm32-wasi/release/spin-js-engine.wasm
 	SPIN_JS_ENGINE_PATH=../../target/wasm32-wasi/release/spin_js_engine.wasm \
 	cargo build --release
 	cp crates/spin-js-engine/src/js_sdk/dist/sdk.d.ts types/lib/index.d.ts
+	cp -r crates/spin-js-engine/src/js_sdk/dist/modules types/lib/
 
 target/wasm32-wasi/release/spin-js-engine.wasm: crates/spin-js-engine/sdk.ts crates/spin-js-engine/src/lib.rs
 	cd crates/spin-js-engine && \
