+++ types/lib/modules/fsPromises.d.ts
@@ -1,0 +1,6 @@
+declare global {
+    const fsPromises: {
+        readFile: (filename: string) => Promise<ArrayBuffer>;
+    };
+}
+export {};
