+++ types/lib/modules/statusTextList.d.ts
@@ -1,0 +1,1 @@
+export {};
