+++ types/lib/modules/stringHandling.d.ts
@@ -1,0 +1,5 @@
+declare global {
+    function atob(data: string): string;
+    function btoa(data: string): string;
+}
+export {};
